+++ Source/DataTransfer.Legacy.MassImport/Core/MassImportManagerNew.cs
@@ -1,6 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using DataTransfer.Legacy.MassImport.RelEyeTelemetry;
+using DataTransfer.Legacy.MassImport.RelEyeTelemetry.Events;
+using DataTransfer.Legacy.MassImport.RelEyeTelemetry.MetricsEventsBuilders;
 using kCura.Utility;
 using Relativity.Core;
 using Relativity.Core.Service;
@@ -63,6 +66,11 @@ namespace Relativity.MassImport.Core
 			}
 
 			var massImportMetric = new MassImportMetrics(CorrelationLogger, APMClient);
+			ITelemetryPublisher telemetryPublisher = new ApmTelemetryPublisher(APMClient);
+			IRelEyeMetricsService relEyeMetricsService = new RelEyeMetricsService(telemetryPublisher);
+			IEventsBuilder eventsBuilder = new EventsBuilder();
+
+
 			var image = new Data.Image(context.DBContext, settings);
 
 			if (!SQLInjectionHelper.IsValidRunId(settings.RunID))
@@ -97,6 +105,8 @@ namespace Relativity.MassImport.Core
 			if (image.IsNewJob)
 			{
 				massImportMetric.SendJobStarted(settings, importType: Constants.ImportType.Images, system: Constants.SystemNames.Kepler);
+				EventJobStart @event = eventsBuilder.BuildJobStartEvent(settings, Constants.ImportType.Images);
+				relEyeMetricsService.PublishEvent(@event);
 			}
 
 			timekeeper.MarkEnd("TempFileInitialization");
@@ -251,6 +261,8 @@ namespace Relativity.MassImport.Core
 					customData: CreateSqlImportMetricsCustomData(settings, retval, image.ImportMeasurements));
 
 				massImportMetric.SendBatchCompleted(settings.RunID, importStopWatch.ElapsedMilliseconds, Constants.ImportType.Images, Constants.SystemNames.Kepler, retval, image.ImportMeasurements);
+				var batchCompletedEvent = eventsBuilder.BuildJobBatchCompletedEvent(retval, Constants.ImportType.Images);
+				relEyeMetricsService.PublishEvent(batchCompletedEvent);
 			}
 
 			retval.RunID = settings.RunID;
@@ -275,6 +287,10 @@ namespace Relativity.MassImport.Core
 			}
 
 			var massImportMetric = new MassImportMetrics(CorrelationLogger, APMClient);
+			ITelemetryPublisher telemetryPublisher = new ApmTelemetryPublisher(APMClient);
+			IRelEyeMetricsService relEyeMetricsService = new RelEyeMetricsService(telemetryPublisher);
+			IEventsBuilder eventsBuilder = new EventsBuilder();
+
 			var productionManager = new Relativity.Core.Service.ProductionManager();
 			var image = new Data.Image(context.DBContext, settings);
 			if (image.HasDataGridWorkToDo && !image.IsDataGridInputValid())
@@ -301,6 +317,8 @@ namespace Relativity.MassImport.Core
 			if (image.IsNewJob)
 			{
 				massImportMetric.SendJobStarted(settings, importType: Constants.ImportType.Production, system: Constants.SystemNames.Kepler);
+				EventJobStart @event = eventsBuilder.BuildJobStartEvent(settings, Constants.ImportType.Production);
+				relEyeMetricsService.PublishEvent(@event);
 			}
 
 			image.ImportMeasurements.StartMeasure(nameof(productionManager.CreateProductionDocumentFileTableForProduction));
@@ -452,6 +470,8 @@ namespace Relativity.MassImport.Core
 					customData: CreateSqlImportMetricsCustomData(settings, retval, image.ImportMeasurements));
 
 				massImportMetric.SendBatchCompleted(settings.RunID, importStopWatch.ElapsedMilliseconds, Constants.ImportType.Production, Constants.SystemNames.Kepler, retval, image.ImportMeasurements);
+				var batchCompletedEvent = eventsBuilder.BuildJobBatchCompletedEvent(retval, Constants.ImportType.Production);
+				relEyeMetricsService.PublishEvent(batchCompletedEvent);
 			}
 
 			retval.RunID = settings.RunID;