+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+## [0.2.65]
+- REL-770037 - Add releye telemetry events
+
+## [0.2.64]
+- REL-790064 - Add missing CAL library
 
 ## [0.2.63]
 - REL-718348 - Map InsufficientAccessControlListPermissions into Kepler PermissionDeniedException